Move parsing to seperate static class, "hexadecial value" instead of "hexadecimal number", inline TryParse - generally cleaner code
